% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova.pcsslm}
\alias{anova.pcsslm}
\alias{anova.pcsslmlist}
\title{ANOVA for linear models fit using PCSS}
\usage{
\method{anova}{pcsslm}(object, ...)

\method{anova}{pcsslmlist}(object, ..., scale = 0, test = "F")
}
\arguments{
\item{object, ...}{objects of class \code{pcsslm}.}

\item{scale}{numeric. An estimate of the noise variance σ^2. If zero this 
will be estimated from the largest model considered.}

\item{test}{a character string specifying the test statistic to be used. Can 
be one of \code{"F"}, \code{"Chisq"} or \code{"Cp"}, with partial matching 
allowed, or \code{NULL} for no test.}
}
\description{
Compute an analysis of variance table for one or more linear model fitted
using PCSS.
}
\author{
R Core Team and contributors worldwide. Modified by Jack Wolf
}
