% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate_addition.R
\name{approx_addition.covars}
\alias{approx_addition.covars}
\title{Approx the sum of two variables based on summary statistics with covariates.}
\usage{
approx_addition.covars(coefs, covariateMeans, covariateVars, weights = 1,
  respVars, respMeans, n)
}
\arguments{
\item{coefs}{A matrix of coefficients. Each column represents intercept,
and slope coefficients for a linear regression of a phenotype as a function
of a genotype and other covariates.}

\item{covariateMeans}{Vector of means of covariates}

\item{covariateVars}{Variance-covariance matrix of all covariates}

\item{weights}{Vector of weights for each phenotype (defaults to 1)}

\item{respVars}{Variance-covariance matrix of all phenotypes}

\item{respMeans}{Vector of means of all phenotypes}

\item{n}{Sample size}
}
\description{
Approx the sum of two variables based on summary statistics with covariates.
}
