% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate_addition.R
\name{approx_addition}
\alias{approx_addition}
\title{Approximate the sum of two variables based on summary statistics.}
\usage{
approx_addition(slopes, stdErrors, intercepts, weights = 1, varX, n,
  covPhenos)
}
\arguments{
\item{slopes}{Vector of slopes of linear regressions of each
phenotype as a function of the genotype}

\item{stdErrors}{Vector of standard errors for each slope coefficient}

\item{intercepts}{Vector of intercepts of linear regressions of each
phenotype as a function of the genotype}

\item{weights}{vector of weights for each phenotype (defaults to 1)}

\item{n}{Sample size}

\item{Variance-covariance}{matrix of all phenotypes}
}
\description{
Approximate the sum of two variables based on summary statistics.
}
