% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplication_estimation.R
\name{approx_conditional}
\alias{approx_conditional}
\title{Approximate the mean of Y conditional on X}
\usage{
approx_conditional(means, covs, response, n)
}
\arguments{
\item{means}{Vector of the mean of X and the mean of Y}

\item{covs}{Matrix of covariances for X and Y}

\item{response}{Character. If "binary" truncates means to interval [0, 1].
If "continuous" does not restrict.}

\item{n}{Sample size}
}
\value{
A list of length 2 consisting of 2 functions that give the
  estimated conditional mean and conditional variance of Y as a function of X
}
\description{
Approximate the mean of Y conditional on X
}
