% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplication_estimation.R
\name{approx_cov}
\alias{approx_cov}
\title{Approximate the covariance of X and Y*Z as well as the mean and variance of Y * Z}
\usage{
approx_cov(means, covs, predictor_type, response, n, f, ...)
}
\description{
Approximate the covariance of X and Y*Z as well as the mean and variance of Y * Z
}
\examples{
ex_data <- bin_data[c("g", "y4", "y5")]
approx_cov(colMeans(ex_data), cov(ex_data), predictor_type = "discrete",
           response = "binary", n = nrow(ex_data),
           f = function(x0) dbinom(x0, size = 2, prob = mean(ex_data$g) / 2),
           support = 0:2)
with(ex_data, cov(g, y4 * y5))

}
