% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplication_estimation.R
\name{approx_mult_prod}
\alias{approx_mult_prod}
\title{Approximate the covariance of a set of predictors and a product of responses}
\usage{
approx_mult_prod(
  means,
  covs,
  n,
  response,
  predictors,
  responses,
  verbose = FALSE
)
}
\arguments{
\item{means}{a vector of predictor and response means with all response means
at the end of the vector.}

\item{covs}{covariance matrix of all predictors and responses with column
and row order corresponding to the order of \code{means}.}

\item{n}{sample size (an integer).}

\item{response}{a string. Currently supports \code{"binary"} or
\code{"continuous"}.}

\item{predictors, responses}{lists of objects of class \code{predictor} where
each entry corresponds to one predictor/response variable.}

\item{verbose}{logical.}
}
\description{
\code{approx_mult_prod} recursively estimates the covariances and means of a
  set of responses
}
\examples{
# 3 Binary Phenotypes
ex_data <- bin_data[c("g", "x", "y1", "y2", "y3")]
head(ex_data)
means <- colMeans(ex_data)
covs <- cov(ex_data)
n <- nrow(ex_data)
predictors <- list(
  new_predictor_snp(maf = mean(ex_data$g) / 2),
  new_predictor_normal(mean = mean(ex_data$x), sd = sd(ex_data$x))
)
responses <- lapply(means[3:length(means)], new_predictor_binary)

approx_mult_prod(means, covs, n, response = "binary",
  predictors = predictors, responses = responses, verbose = T)
# Compare to
w3 <- with(ex_data, y1 * y2 * y3)
colMeans(cbind(ex_data[1:2], w3))
cov(cbind(ex_data[1:2], w3))

}
