% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical_estimation.R
\name{approx_or_2_OLD}
\alias{approx_or_2_OLD}
\title{Approximate a linear model for a logical OR statement with 2 disjuncts}
\usage{
approx_or_2_OLD(
  means,
  covs,
  n,
  predictors,
  verbose = FALSE,
  add_intercept = TRUE
)
}
\arguments{
\item{means}{vector of predictor and response means with the last two means
the means of two binary responses to combine in a logical or statement}

\item{covs}{a matrix of the covariance of all model predictors and the
responses with the order of rows/columns corresponding to the order of
\code{means}.}

\item{n}{sample size.}

\item{predictors}{list of objects of class \code{predictor} corresponding
to the order of the predictors in \code{means}.}

\item{add_intercept}{logical. Should the linear model add an intercept term?}
}
\description{
\code{approx_or_2} approximates the linear model for the response
  "y1 or y2" as a function of a set of predictors.
}
