% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_pca.R
\name{approx_pca}
\alias{approx_pca}
\title{Approximate the relationship between a genotype and
the principal components of a set of phenotypes.}
\usage{
approx_pca(slopes, stdErrors, intercepts, varX, covPhenos, n)
}
\arguments{
\item{slopes}{Vector of slopes of linear regressions of each
phenotype as a function of the genotype}

\item{stdErrors}{Vector of standard errors for each slope coefficient}

\item{intercepts}{Vector of intercepts of linear regressions of each
phenotype as a function of the genotype}

\item{varX}{Variance of X, the independent variable (genotype)}

\item{covPhenos}{(Optional) Variance-covariance matrix of all phenotypes}

\item{n}{Sample size}
}
\description{
Approximate the relationship between a genotype and
the principal components of a set of phenotypes.
}
