% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplication_estimation.R
\name{approx_prod_recursive}
\alias{approx_prod_recursive}
\title{Recursively approximate covariances with a product of responses}
\usage{
approx_prod_recursive(means, covs, n, response, predictors, responses)
}
\description{
Recursively approximate covariances with a product of responses
}
