% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplication_estimation.R
\name{approx_prod_stats}
\alias{approx_prod_stats}
\title{Approximate summary statistics for a product of phenotypes and a set of predictors}
\usage{
approx_prod_stats(means, covs, n, response, predictors)
}
\arguments{
\item{means}{Vector of means of predictors and the two phenotypes to be multiplied}

\item{covs}{Covariance matrix of all predictors and the two phenotypes}

\item{n}{Sample size}

\item{predictors}{a list of elements of class predictor}
}
\value{
A list with the predicted covariance matrix of all predictors and
  the product and the means of all predictors and the product.
}
\description{
Approximate summary statistics for a product of phenotypes and a set of predictors
}
\examples{
ex_data <- cont_data[c("g", "x", "y3", "y5")]
predictors <- list(
  new_predictor_snp(maf = mean(ex_data$g) / 2),
  new_predictor_normal(mean = mean(ex_data$x), sd = sd(ex_data$x))
  )
approx_prod_stats(colMeans(ex_data), covs = cov(ex_data), n = nrow(ex_data),
  response = "continuous", predictors = predictors)
with(ex_data, cov(cbind(g, x, y3y5 = y3 * y5)))

}
