% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplication_estimation.R
\name{approx_response_cov_recursive}
\alias{approx_response_cov_recursive}
\title{Approximate the covariance of one response with an arbitrary product of
responses.}
\usage{
approx_response_cov_recursive(
  ids,
  r_covs,
  r_means,
  n,
  responses,
  response,
  verbose = FALSE
)
}
\arguments{
\item{ids}{Column ids of responses to use. First is taken alone while 2nd to
last are to be multiplied}

\item{r_covs}{Response covariance matrix}

\item{r_means}{Response means (vector)}

\item{n}{Sample size}

\item{responses}{List of lists with elements of class predictor}

\item{response}{Character, Either "binary" or "continuous"}
}
\value{
A vector with the approximated covariance, and approximated mean and
  variance of the product
}
\description{
Approximate the covariance of one response with an arbitrary product of
responses.
}
\examples{
Y <- bin_data[c("y1", "y2", "y3")]
r_means <- colMeans(Y)
r_covs <- cov(Y)
n <- nrow(Y)
responses <- lapply(r_means, new_predictor_binary)
approx_response_cov_recursive(ids = c(3, 2, 1), r_covs, r_means, n, responses, response = "binary", verbose = T)
w2 <- with(Y, y1 * y2)
cov(Y$y3, w2)
mean(w2)
var(w2)


}
