% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dichotomize.R
\name{approxDichotomous}
\alias{approxDichotomous}
\title{Approximate coefficients for a dichotomous version of a continuous variable}
\usage{
approxDichotomous(slope, intercept, varY, varX, maf, cutoff, n,
  inheritance)
}
\arguments{
\item{slope}{Slope coefficient for the original continuous response as
a function of the genoptype}

\item{intercept}{Intercept coefficient for the original continuous
response as a function of the genoptype}

\item{varY}{Variance of the original continuous response}

\item{varX}{Variance of the genotype}

\item{maf}{Minor allele frequency}

\item{cutoff}{Cutoff point for dichotomization}

\item{n}{Sample size}
}
\description{
'\code{dichotmize} approximates the relationship between a genotype and
a dichotomous version of a phenotype given information about the relation
between the genotype and a linear version of the phenotype. E.g. given
information about the relationship between X and blood pressure, approximate
the model between X and wether a patient has high blood pressure
(BP > 140mm Hg)
}
