% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplication_estimators.R
\name{approxPhenoProd}
\alias{approxPhenoProd}
\title{Approximate coefficients for a product of phenotypes}
\usage{
approxPhenoProd(slopes, intercepts, maf, inheritance = "dominant",
  varCovar, yBar, n)
}
\arguments{
\item{slopes}{A vector of slope coeffients for each phenotype as a function
of a common genotype}

\item{intercepts}{A vector of intercepts for each phenotype as a function of
a common genotype}

\item{maf}{Minor allele frequency}

\item{inheritance}{Phenotype inheritance method. Either 'domiant' or
'additive'}

\item{varCovar}{Variance-covariance matrix of all phenotypes}

\item{yBar}{Vector of means of phenotypes}

\item{n}{Sample size}
}
\description{
Approximate coefficients for a product of phenotypes
}
\seealso{
\code{\link{getMultCoefs}} for coeffient estimation and
 \code{\link{getMultSE}} for coefficient standard error estimation.
}
