% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bin_data}
\alias{bin_data}
\title{Simulated data with binary phenotypes}
\format{
A data frame with 1000 rows and 7 columns:
\describe{
  \item{g}{Minor allele counts at one site}
  \item{x}{A continuous covariate}
  \item{y1,y2,y3,y4,y5}{binary phenotypes}
}
}
\usage{
bin_data
}
\description{
A dataset containing simulated genetic data with one SNP, a covariate, and
5 binary phenotypes.
}
\keyword{datasets}
