% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{check_terms}
\alias{check_terms}
\title{Check that independent and dependent variables are accounted for through PCSS}
\usage{
check_terms(xterms, yterms, pcssterms, pcsstype)
}
\arguments{
\item{xterms, }{yterms character vector of model's independent variables or
variables combined to the dependent variable}

\item{pcssterms}{character vector of variables with provided PCSS}

\item{pcsstype}{character describing the PCSS being checked. Either
\code{"means"}, \code{"covs"}, \code{"predictors"}, or
\code{"responses"}.}
}
\description{
Check that independent and dependent variables are accounted for through PCSS
}
