% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cont_data}
\alias{cont_data}
\title{Simulated data with continuous phenotypes}
\format{
A data frame with 1000 rows and 7 columns:
\describe{
  \item{g}{Minor allele counts at one site}
  \item{x}{A continuous covariate}
  \item{y1,y2,y3,y4,y5}{5 continuous phenotypes}
}
}
\usage{
cont_data
}
\description{
A dataset containing simulated genetic data with one SNP, a covariate, and
5 continuous phenotypes.
}
\keyword{datasets}
