% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplication_estimation.R
\name{get_mean}
\alias{get_mean}
\title{Calculate the mean of the product of X and Y}
\usage{
get_mean(means, covs, n)
}
\description{
Calculate the mean of the product of X and Y
}
\examples{
n <- 100
x <- rnorm(n = n, mean = 0, sd = 1)
y <- rnorm(n = n, mean = 2 - 0.01 * x, sd = 2)
XY <- cbind(x, y)
get_mean(colMeans(XY), cov(XY), n)
mean(x * y)
}
