% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatReg.R
\name{getCoefs.grass}
\alias{getCoefs.grass}
\title{Approximate coefficients for a product of phenotypes}
\usage{
getCoefs.grass(slopes, intercepts, maf = NULL,
  inheritance = "dominant")
}
\arguments{
\item{slopes}{A vector of slope coeffients for each phenotype as a function
of a common genotype}

\item{intercepts}{A vector of intercepts for each phenotype as a function of
a common genotype}

\item{maf}{Minor allele frequency. Required for inheritance = 'additive'}

\item{inheritance}{Phenotype inheritance method. Either 'domiant' or
'additive'}
}
\description{
Approximate coefficients for a product of phenotypes
}
