% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplication_estimators.R
\name{getMultSE}
\alias{getMultSE}
\title{Approximate slope standard error for a product of phenotypes}
\usage{
getMultSE(slopes, intercepts, maf, varCovar, yBar, n, predBeta,
  inheritance)
}
\arguments{
\item{slopes}{A vector of slope coeffients for each phenotype as a function
of a common genotype}

\item{intercepts}{A vector of intercepts for each phenotype as a function of
a common genotype}

\item{maf}{Minor allele frequency}

\item{varCovar}{Variance-covariance matrix of all phenotypes}

\item{yBar}{Vector of means of phenotypes}

\item{n}{Sample size}

\item{predBeta}{Predicted slope from getCoefs()}

\item{inheritance}{Phenotype inheritance method. Either 'domiant' or
'additive'}
}
\description{
Approximate slope standard error for a product of phenotypes
}
