% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{guess_response}
\alias{guess_response}
\title{Guess the function that is applied to a set of responses}
\usage{
guess_response(response = character())
}
\arguments{
\item{response}{character. Output of \code{extract_response}.}
}
\value{
A character. Either \code{"model_combo"}, \code{"model_product"}, 
\code{"model_or"}, \code{"model_and"}, or \code{"model_singular"}.
}
\description{
\code{guess_response} takes a character vector of the dependent variable
from a \code{formula} object and identifies which function separates the
individual variables that make up the response. It then returns the
\code{model_*} function to model the appropriate response using PCSS.
}
