% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addition.R
\name{lm_adjust}
\alias{lm_adjust}
\title{Adjust a linear model for covariates}
\usage{
lm_adjust(coefs, xMeans, xCov, yMean, yVar, n)
}
\arguments{
\item{coefs}{2 x p Matrix of model coefficients for y ~ x_1, ... y ~ x_p
The first and second rows correspond to intercept and slope coefficients
for y ~ x, respectively.}

\item{xMeans}{Vector of length p of means of all covariates}

\item{xCov}{Covariance matrix of all covariates}

\item{yMean}{Response mean}

\item{yVar}{Response variance}

\item{n}{Sample size}
}
\description{
Adjust a linear model for covariates
}
\examples{
data(mtcars)

dispMod <- lm(mpg ~ disp, data = mtcars)
cylMod  <- lm(mpg ~ cyl,  data = mtcars)

coefs <- cbind(coefficients(dispMod), coefficients(cylMod))

xMeans <- c(mean(mtcars$disp), mean(mtcars$cyl))

xCov <- cov(mtcars)[c('disp', 'cyl'), c('disp', 'cyl')]

yMean <- mean(mtcars$mpg)
yVar <- var(mtcars$mpg)

n <- nrow(mtcars)

out <- lm_adjust(coefs = coefs,
                 xMeans = xMeans,
                 xCov = xCov,
                 yMean = yMean,
                 yVar = yVar,
                 n = n)

out
summary(lm(mpg ~ disp + cyl, data = mtcars))$coefficients
}
