% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addition.R
\name{lm_combine}
\alias{lm_combine}
\title{Model a linear combination of phenotypes}
\usage{
lm_combine(coefs, xMeans, xCov, yMeans, yCov, n, weights = 1)
}
\arguments{
\item{coefs}{(p + 1) * m matrix of model coefficients for
y_1 ~ x_1 + ... + x_p, ..., y_m ~ x_1 + ... + x_p.
Columns correspond to responses and rows correspond to covariates.}

\item{xMeans}{Vector of length p of all means of covariates}

\item{xCov}{Covariance matrix of all covariates}

\item{yMeans}{Vector of length m of all means of responses}

\item{yCov}{Covariance matrix of all responses}

\item{n}{Sample size}

\item{weights}{Vector of length m of phenotype weights. Defaults to
equal weights (c(1, ... ,1)).}
}
\description{
Model a linear combination of phenotypes
}
\examples{
mpgMod <- lm(mpg ~ disp + cyl, data = mtcars)
hpMod  <- lm(hp  ~ disp + cyl, data = mtcars)

coefs <- cbind(coefficients(mpgMod), coefficients(hpMod))
xMeans <- c(mean(mtcars$disp), mean(mtcars$cyl))
xCov <- cov(mtcars)[c('disp', 'cyl'), c('disp', 'cyl')]

yMeans <- c(mean(mtcars$mpg), mean(mtcars$hp))
yCov   <- cov(mtcars)[c('mpg', 'hp'), c('mpg', 'hp')]

n <- nrow(mtcars)

out <- lm_combine(coefs = coefs,
                  xMeans = xMeans,
                  xCov = xCov,
                  yMeans = yMeans,
                  yCov = yCov,
                  n = n,
                  weights = 1)

out
summary(lm((mtcars$mpg + mtcars$hp) ~ disp + cyl, data = mtcars))$coefficients
}
