% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplication_estimation.R
\name{make_permutations}
\alias{make_permutations}
\title{List all permutations of a sequence of integers}
\usage{
make_permutations(m)
}
\arguments{
\item{m}{number of elements to permute}
}
\description{
Lists all permutations of 1,2,...,m unique up to the first two elements
}
