% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{meta_pcsslm}
\alias{meta_pcsslm}
\title{Approximate a linear model using PCSS from multiple sources}
\usage{
meta_pcsslm(formula, pcss_list = list(), ...)
}
\arguments{
\item{formula}{an object of class formula whose dependent variable is a 
combination of variables and logical | operators. 
All model terms must have appropriate PCSS in \code{pcss}.}

\item{pcss_list}{A named list of lists of precomputed summary statistics.
Each element should correspond to a study or source and match the format
described for argument \code{pcss} in \code{\link{pcsslm}}. Names,
content, and orderings must be consistent across studies.}

\item{...}{additional arguments. See Details for more information.}
}
\description{
\code{meta_pcsslm} pools summary statistics across sources and then 
  approximates a lineaer model using the pooled summary statistics as input.
}
\note{
\code{meta_pcsslm} currently does not support models that require
  response multiplication including products, or logical disjunctions or
  conjunctions.
}
\examples{
## Principal Component Analysis
data("pcsstools_example")
ex_data <- pcsstools_example[c("g1", "g2", "y1", "y2", "Source")]
pcss_list <- list(
  "Study_1" = list(
    means = colMeans(ex_data[ex_data$Source == 1, ]),
    covs = cov(ex_data[ex_data$Source == 1, ]),
    n = nrow(ex_data[ex_data$Source == 1, ])
  ),
  
  "Study_2" = list(
    means = colMeans(ex_data[ex_data$Source == 2, ]),
    covs = cov(ex_data[ex_data$Source == 2, ]),
    n = nrow(ex_data[ex_data$Source == 2, ])
  ),
  
  "Study_3" = list(
    means = colMeans(ex_data[ex_data$Source == 3, ]),
    covs = cov(ex_data[ex_data$Source == 3, ]),
    n = nrow(ex_data[ex_data$Source == 3, ])
  )
)

meta_pcsslm(y1 + y2 + y3 ~ g1 + x1, pcss_list = pcss_list, comp = 1)

}
