% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical_estimation.R
\name{model_and}
\alias{model_and}
\title{Approximate a linear model for a series of logical AND statements using PCSS}
\usage{
model_and(formula, n, means, covs, predictors, ...)
}
\arguments{
\item{formula}{an object of class \code{formula} whose dependent variable is
a combination of variables and logical \code{&} operators. All model terms
must be accounted for in \code{means} and \code{covs}.}

\item{n}{sample size.}

\item{means}{named vector of predictor and response means.}

\item{covs}{named matrix of the covariance of all model predictors and the
responses.}

\item{predictors}{named list of objects of class \code{predictor}.}

\item{...}{additional arguments}
}
\description{
\code{model_and} approximates the linear model for the a conjunction
  of m phenotypes as a function of a set of predictors.
}
\examples{
ex_data <- bin_data[c("g", "x", "y1", "y2")]
head(ex_data)
means <- colMeans(ex_data)
covs <- cov(ex_data)
n <- nrow(ex_data)
predictors <- list(
  g = new_predictor_snp(maf = mean(ex_data$g) / 2),
  x = new_predictor_normal(mean = mean(ex_data$x), sd = sd(ex_data$x))
)

model_and(
  y1 & y2 ~ g + x, means = means, covs = covs, n = n, predictors = predictors
)
summary(lm(y1 & y2 ~ g + x + 1, data = ex_data))

}
