% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_linear_combo.R
\name{model_singular}
\alias{model_singular}
\title{Model an individual phenotype using PCSS}
\usage{
model_singular(formula, n, means, covs, ...)
}
\arguments{
\item{formula}{an object of class \code{formula} whose dependent variable is
only variable. All model terms must be accounted for in \code{means}
and \code{covs}.}

\item{n}{sample size.}

\item{means}{named vector of predictor and response means.}

\item{covs}{named matrix of the covariance of all model predictors and the
responses.}

\item{...}{additional arguments}
}
\description{
\code{model_singular} calculates the linear model for a singular
  phenotype as a function of a set of predictors.
}
\examples{
ex_data <- cont_data[c("g", "x", "y1")]
means <- colMeans(ex_data)
covs <- cov(ex_data)
n <- nrow(ex_data)

model_singular(
  y1 ~ g + x,
  n = n, means = means, covs = covs
)
summary(lm(y1 ~ g + x, data = ex_data))
}
\references{
Wolf, J.M., Barnard, M., Xueting, X., Ryder, N., Westra, J., and
  Tintle, N.  (2020). Computationally efficient, exact, covariate-adjusted
  genetic principal component analysis by leveraging individual marker
  summary statistics from large biobanks. \emph{Pacific Symposium on
  Biocomputing}, 25, 719-730.
}
