% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor_class.R
\name{new_predictor_binary}
\alias{new_predictor_binary}
\title{Shortcut to create a predictor object for a binary variable}
\usage{
new_predictor_binary(p)
}
\arguments{
\item{p}{probability of success (predictor mean)}
}
\description{
\code{new_predictor_binary} calls \code{new_predictor}
}
\examples{
new_predictor_binary(p = 0.75)

}
