% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor_class.R
\name{new_predictor_normal}
\alias{new_predictor_normal}
\title{Shortcut to create a predictor object for a continuous variable}
\usage{
new_predictor_normal(mean, sd)
}
\arguments{
\item{mean}{predictor mean (double).}

\item{sd}{predictor standard deviation (double)}
}
\description{
\code{new_predictor_normal} calls \code{new_predictor}
}
\examples{
new_predictor_normal(mean = 10, sd = 1)

}
