% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor_class.R
\name{new_predictor_snp}
\alias{new_predictor_snp}
\title{Shortcut to create a predictor object for a SNP's minor allele counts}
\usage{
new_predictor_snp(maf)
}
\arguments{
\item{maf}{minor allele frequency}
}
\value{
an object of class predictor.
}
\description{
\code{new_predictor_snp} calls \code{new_predictor}
}
\examples{
new_predictor_snp(maf = 0.3)
}
