% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcsstools.R
\docType{package}
\name{pcsstools}
\alias{pcsstools}
\title{pcsstools: Tools for Regression Using Pre-Computed Summary Statistics}
\description{
The pcsstools package provides functions to describe various regression 
models using only pre-computed summary statistics (PCSS) from genome-wide 
association studies (GWASs) and PCSS repositories.
}
\details{
The main function of interest is \code{\link{pcsslm}}, which can be used to
approximate linear models for various combinations of phenotypes using
PCSS.
}
