% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcsstools_example}
\alias{pcsstools_example}
\title{Simulated example data}
\format{
A data frame with 1000 rows and 13 columns:
\describe{
  \item{g1,g2,g3}{Minor allele counts at three sites}
  \item{x1,x2,x3}{Continuous covariates}
  \item{y1,y2,y3}{Continuous phenotypes}
  \item{y4,y5,y6}{Binary phenotypes}
  \item{Source}{Source indicator}
}
}
\usage{
pcsstools_example
}
\description{
A dataset containing simulated genetic data with 3 SNPs, 3 continuous 
covariates, 6 continuous phenotypes, and a source indicator.
}
\keyword{datasets}
