% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{pool_meta_stats}
\alias{pool_meta_stats}
\title{Aggregate PCSS from multiple studies into PCSS for pooled data}
\usage{
pool_meta_stats(pcss_list)
}
\arguments{
\item{pcss_list}{A named list of lists of precomputed summary statistics.
Each element should correspond to a study or source and match the format
described for argument \code{pcss} in \code{\link{pcsslm}}. Names,
content, and orderings must be consistent across studies.}
}
\value{
A list of precomputed summary statistics for the pooled data of
  the format described for argument \code{pcss} in \code{\link{pcsslm}}.
}
\description{
Aggregate PCSS from multiple studies into PCSS for pooled data
}
\author{
Nathan Ryder
}
