% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{pool_predictors}
\alias{pool_predictors}
\title{Aggregate predictors from multiple studies}
\usage{
pool_predictors(pcss_list)
}
\arguments{
\item{pcss_list}{A named list of lists of precomputed summary statistics.
Each element should correspond to a study or source and match the format
described for argument \code{pcss} in \code{\link{pcsslm}}. Names,
content, and orderings must be consistent across studies.}
}
\description{
Aggregate predictors from multiple studies
}
