% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_pcss_class.R
\name{print.summary.lm_pcss}
\alias{print.summary.lm_pcss}
\title{Print an object of class summary.lm_pcss}
\usage{
\method{print}{summary.lm_pcss}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  symbolic.cor = x$symbolic.cor,
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{x}{an object of class \code{"summary.lm_pcss"}}

\item{digits}{the number of significant digits to use when printing.}

\item{symbolic.cor}{logical. If TRUE, print the correlations in a symbolic
form (see \link[stats]{symnum}) rather than as numbers.}

\item{signif.stars}{logical. If \code{TRUE}, ‘significance stars’ are printed
for each coefficient.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Prints a linear model fit through pre-computed summary statistics
}
\author{
R Core Team and contributors worldwide. Modified by Jack Wolf
}
