% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_models.R
\name{calculate_lm}
\alias{calculate_lm}
\title{Calculate a linear model using PCSS}
\usage{
calculate_lm(
  means,
  covs,
  n,
  add_intercept = FALSE,
  keep_pcss = FALSE,
  terms = NULL
)
}
\arguments{
\item{means}{a vector of means of all model predictors and the response with
the last element the response mean.}

\item{covs}{a matrix of the covariance of all model predictors and the
response with the order of rows/columns corresponding to the order of
\code{means}.}

\item{n}{sample size}

\item{add_intercept}{logical. If \code{TRUE} adds an intercept to the model.}

\item{keep_pcss}{logical. If \code{TRUE}, returns \code{means} and 
\code{covs}.}

\item{terms}{terms}
}
\description{
\code{calculate_lm} describes the linear model of the last listed variable
in \code{means} and \code{covs} as a function of all other variables in
\code{means} and \code{covs}.
}
\references{
{

  \insertRef{wolf_using_2021}{pcsstools}
  
  \insertRef{wolf_computationally_2020}{pcsstools}
  
  \insertRef{gasdaska_leveraging_2019}{pcsstools}

}
}
