% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_models.R
\name{calculate_lm}
\alias{calculate_lm}
\title{Calculate a linear model using PCSS}
\usage{
calculate_lm(
  means,
  covs,
  n,
  add_intercept = FALSE,
  keep_pcss = FALSE,
  terms = NULL
)
}
\arguments{
\item{means}{a vector of means of all model predictors and the response with
the last element the response mean.}

\item{covs}{a matrix of the covariance of all model predictors and the
response with the order of rows/columns corresponding to the order of
\code{means}.}

\item{n}{sample size}

\item{add_intercept}{logical. If \code{TRUE} adds an intercept to the model.}

\item{keep_pcss}{logical. If \code{TRUE}, returns \code{means} and 
\code{covs}.}

\item{terms}{terms}
}
\description{
\code{calculate_lm} describes the linear model of the last listed variable
in \code{means} and \code{covs} as a function of all other variables in
\code{means} and \code{covs}.
}
\references{
Wolf, J.M., Barnard, M., Xueting, X., Ryder, N., Westra, J., and
  Tintle, N.  (2020). Computationally efficient, exact, covariate-adjusted
  genetic principal component analysis by leveraging individual marker
  summary statistics from large biobanks. \emph{Pacific Symposium on
  Biocomputing}, 25, 719-730.

  Gasdaska A., Friend D., Chen R., Westra J., Zawistowski M., Lindsey W. and
  Tintle N. (2019) Leveraging summary statistics to make inferences about
  complex phenotypes in large biobanks. \emph{Pacific Symposium on
  Biocomputing}, 24, 391-402.
}
