% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_linear_combo.R
\name{calculate_lm_combo}
\alias{calculate_lm_combo}
\title{Calculate a linear model for a linear combination of responses}
\usage{
calculate_lm_combo(means, covs, n, phi, m = length(phi), add_intercept, ...)
}
\arguments{
\item{means}{a vector of means of all model predictors and the response with
the last \code{m} elements the response means (with order corresponding to
the order of weights in \code{phi}).}

\item{covs}{a matrix of the covariance of all model predictors and the
responses with the order of rows/columns corresponding to the order of
\code{means}.}

\item{n}{sample size.}

\item{phi}{vector of linear combination weights with one entry per response
variable.}

\item{m}{number of responses to combine. Defaults to \code{length(weighs)}.}

\item{add_intercept}{logical. If \code{TRUE} adds an intercept to the model.}

\item{...}{additional arguments}
}
\description{
\code{calculate_lm_combo} describes the linear model for a linear combination
  of responses as a function of a set of predictors.
}
\references{
Wolf, J.M., Barnard, M., Xueting, X., Ryder, N., Westra, J., and
  Tintle, N.  (2020). Computationally efficient, exact, covariate-adjusted
  genetic principal component analysis by leveraging individual marker
  summary statistics from large biobanks. \emph{Pacific Symposium on
  Biocomputing}, 25, 719-730.

  Gasdaska A., Friend D., Chen R., Westra J., Zawistowski M., Lindsey W. and
  Tintle N. (2019) Leveraging summary statistics to make inferences about
  complex phenotypes in large biobanks. \emph{Pacific Symposium on
  Biocomputing}, 24, 391-402.
}
