% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{extract_predictors}
\alias{extract_predictors}
\title{Extract independent variables from a formula}
\usage{
extract_predictors(formula = formula())
}
\arguments{
\item{formula}{an object of class \code{formula}.}

\item{all_vars}{character vector of possible terms}
}
\description{
Extract independent variables from a formula
}
