% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{extract_response}
\alias{extract_response}
\title{Extract dependent variables from a formula as a string}
\usage{
extract_response(formula = formula())
}
\arguments{
\item{formula}{an object of class \code{formula}.}
}
\value{
a character vector of all responses
}
\description{
Extract dependent variables from a formula as a string
}
