% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_linear_combo.R
\name{model_prcomp}
\alias{model_prcomp}
\title{Model the principal component score of a set of phenotypes using PCSS}
\usage{
model_prcomp(formula, comp = 1, n, means, covs, ...)
}
\arguments{
\item{formula}{an object of class \code{formula} whose dependent variable is
a series of variables joined by \code{+} operators. \code{model_prcomp}
will treat a principal component score of those variables as the actual
dependent variable. All model terms must be accounted for in \code{means} 
and \code{covs}.}

\item{comp}{integer indicating which principal component score to analyze.
Must be less than or equal to the total number of phenotypes.}

\item{n}{sample size.}

\item{means}{named vector of predictor and response means.}

\item{covs}{named matrix of the covariance of all model predictors and the
responses.}

\item{...}{additional arguments}
}
\description{
\code{model_prcomp} calculates the linear model for the mth principal 
  component score of a set of phenotypes as a function of a set of 
  predictors.
}
\examples{
ex_data <- cont_data[c("g", "x", "y1", "y2", "y3")]
head(ex_data)
means <- colMeans(ex_data)
covs <- cov(ex_data)
n <- nrow(ex_data)   

model_prcomp(
  y1 + y2 + y3 ~ g + x, comp = 1, n = n, means = means, covs = covs
)


}
\references{
Wolf, J.M., Barnard, M., Xueting, X., Ryder, N., Westra, J., and 
  Tintle, N.  (2020). Computationally efficient, exact, covariate-adjusted 
  genetic principal component analysis by leveraging individual marker 
  summary statistics from large biobanks. \emph{Pacific Symposium on 
  Biocomputing}, 25, 719-730.

  Gasdaska A., Friend D., Chen R., Westra J., Zawistowski M., Lindsey W. and 
  Tintle N. (2019) Leveraging summary statistics to make inferences about 
  complex phenotypes in large biobanks. \emph{Pacific Symposium on 
  Biocomputing}, 24, 391-402.
}
